Rename tabulation open_primary module to combined_ballot_primary
Renames the module file and its import paths; exported helpers are unchanged.

Part of the open-primary -> combined ballot primary rename (#8769).
