+++ libs/utils/src/tabulation/index.ts
@@ -4,7 +4,7 @@ export * from './convert';
 export * from './compressed_tallies';
 export * as CachedElectionLookups from './lookups';
 export * from './mock_tally_report_results';
-export * from './open_primary';
+export * from './combined_ballot_primary';
 export * from './straight_party';
 export * from './tabulation';
 export * from './tally_reports';
